backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["numbersflow <eriko20190801@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = "^0.31.1"
requests = "^2.32.3"
httpx = "^0.27.2"
python-dotenv = "^1.0.1"
pytz = "^2024.2"
pyyaml = "^6.0.2"
redis = "^5.1.1"
psutil = "^6.0.0"
gputil = "^1.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
